Update ADD_MESSAGE to be burstified
