+++ src/client/components/chat/ChatContent.spec.js
@@ -6,32 +6,29 @@ import ChatContent from './ChatContent.jsx';
 
 const setup = () => {
   const props = {
-    bursts: {
-      'liz:2017-01-31T23:22:38.808Z': [
-        {
-          _id: '58911c3e871fdf24f2079782',
-          createdAt: '2017-01-31T23:22:38.808Z',
-          username: 'liz',
-          text: 'sweet',
-          burstStart: true
-        }
-      ],
-      'dan:2017-02-01T05:34:40.591Z': [
-        {
-          _id: '58917370eb323841b839f583',
-          createdAt: '2017-02-01T05:34:40.591Z',
-          username: 'dan',
-          text: 'cool',
-          burstStart: true
-        },
-        {
-          _id: '5891737ceb323841b839f584',
-          createdAt: '2017-02-01T05:34:52.011Z',
-          username: 'dan',
-          text: 'damn it',
-          burstStart: false
-        }
-      ]
+    messages: {
+      '58911c3e871fdf24f2079782': {
+        _id: '58911c3e871fdf24f2079782',
+        createdAt: '2017-01-31T23:22:38.808Z',
+        username: 'liz',
+        text: 'sweet',
+        burstStart: true
+      },
+      '58917370eb323841b839f583': {
+        _id: '58917370eb323841b839f583',
+        createdAt: '2017-02-01T05:34:40.591Z',
+        username: 'dan',
+        text: 'cool',
+        burstStart: true
+      },
+
+      '5891737ceb323841b839f584': {
+        _id: '5891737ceb323841b839f584',
+        createdAt: '2017-02-01T05:34:52.011Z',
+        username: 'dan',
+        text: 'damn it',
+        burstStart: false
+      }
     }
   };
   const component = <ChatContent {...props} />;
