+++ src/client/reducers/messages.spec.js
@@ -12,12 +12,17 @@ describe('messages reducer', () => {
       messages({}, {
         type: 'ADD_MESSAGE',
         message: {
+          createdAt: '2017-02-01T09:00:00-08:00',
+          username: 'dan',
           text: 'Run the tests',
           _id: 0
         }
       })
     ).toEqual({
       0: {
+        burstStart: true,
+        createdAt: '2017-02-01T09:00:00-08:00',
+        username: 'dan',
         text: 'Run the tests',
         _id: 0
       }
@@ -26,22 +31,33 @@ describe('messages reducer', () => {
     expect(
       messages({
         0: {
+          createdAt: '2017-02-01T09:00:00-08:00',
+          username: 'dan',
           text: 'Run the tests',
           _id: 0
         }
       }, {
         type: 'ADD_MESSAGE',
         message: {
+          // 1 min later, same burst
+          createdAt: '2017-02-01T09:01:00-08:00',
+          username: 'dan',
           text: 'Use Redux',
           _id: 1
         }
       })
     ).toEqual({
       0: {
+        createdAt: '2017-02-01T09:00:00-08:00',
+        username: 'dan',
+        burstStart: true,
         text: 'Run the tests',
         _id: 0
       },
       1: {
+        createdAt: '2017-02-01T09:01:00-08:00',
+        username: 'dan',
+        burstStart: false,
         text: 'Use Redux',
         _id: 1
       }
@@ -50,30 +66,47 @@ describe('messages reducer', () => {
     expect(
       messages({
         0: {
+          createdAt: '2017-02-01T09:00:00-08:00',
+          username: 'dan',
           text: 'Run the tests',
           _id: 0
         },
         1: {
+          // 1 min later, same burst
+          createdAt: '2017-02-01T09:01:00-08:00',
+          username: 'dan',
           text: 'Use Redux',
           _id: 1
         }
       }, {
         type: 'ADD_MESSAGE',
         message: {
+           // 20 min later, new burst
+          createdAt: '2017-02-01T09:20:00-08:00',
+          username: 'dan',
           text: 'Fix the tests',
           _id: 2
         }
       })
     ).toEqual({
       0: {
+        createdAt: '2017-02-01T09:00:00-08:00',
+        username: 'dan',
+        burstStart: true,
         text: 'Run the tests',
         _id: 0
       },
       1: {
+        createdAt: '2017-02-01T09:01:00-08:00',
+        username: 'dan',
+        burstStart: false,
         text: 'Use Redux',
         _id: 1
       },
       2: {
+        createdAt: '2017-02-01T09:20:00-08:00',
+        username: 'dan',
+        burstStart: true,
         text: 'Fix the tests',
         _id: 2
       }
